refactored landing page file structure
